# Add performance report