Example:
```
Report for 2023-04-07 18:18:14 run
==================================
Legend
======
   impl_postfix                                    description
0  numba_dpex_k                         Numba-dpex with kernel
1  numba_dpex_p                         Numba-dpex with prange
2  numba_dpex_n         Numba-dpex with NumPy array expression
3          dpnp                   Data-parallel Numeric Python
4         numpy                                          Numpy
5        python                                         Python
6       numba_n                            Numba with nopython
7      numba_np             Numba with nopython, Parallel=True
8     numba_npr  Numba with nopython, Parallel=True and prange
9          sycl                      DPC++ with dpctl bindings

Summary of current performance
=================================
  input_size          benchmark problem_preset     numba_dpex_k     numba_dpex_p   numba_dpex_n                dpnp       numpy           python         numba_n        numba_np      numba_npr            sycl
0       15MB      black_scholes              S    0.03s, 14.98%    0.01s, -33.8%  0.04s, 94.62%      0.11s, 393.08%  0.02s, ref  0.36s, 1549.92%  0.02s, -29.47%   0.0s, -94.47%  0.0s, -94.57%   0.0s, -95.72%
1       80KB             dbscan              S   0.08s, 972.35%    0.02s, 134.5%            n/a                 n/a         n/a       0.01s, ref   0.01s, 61.86%             n/a  0.0s, -60.48%   0.0s, -63.67%
2        4KB             gpairs              S  0.05s, 7767.25%  0.03s, 5088.03%            n/a    0.09s, 13422.42%   0.0s, ref              n/a   0.0s, -52.48%             n/a  0.0s, -74.58%    0.0s, -56.0%
3       94KB             kmeans              S   0.73s, 275.32%   0.59s, 202.94%            n/a                 n/a         n/a       0.19s, ref   0.0s, -99.27%             n/a  0.0s, -99.32%  0.01s, -97.35%
4      296KB                knn              S    0.03s, -99.0%              n/a            n/a                 n/a         n/a       3.06s, ref             n/a             n/a  0.0s, -99.92%   0.0s, -99.93%
5        1MB            l2_norm              S  0.01s, 3410.52%  0.01s, 2922.57%            n/a  8.66s, 2780288.33%   0.0s, ref              n/a   0.0s, -23.67%   0.0s, -36.36%  0.0s, -80.97%   0.0s, -26.95%
6        8MB  pairwise_distance              S    0.02s, 62.59%     0.01s, 3.76%            n/a                 n/a  0.01s, ref              n/a             n/a    0.02s, 82.4%  0.0s, -91.18%   0.0s, -92.13%
7         0B                pca              S              n/a              n/a            n/a                 n/a  0.03s, ref              n/a  0.07s, 111.03%  0.02s, -22.24%            n/a             n/a
8        7MB              rambo              S   0.03s, 443.15%   0.01s, 159.89%            n/a      0.04s, 641.58%  0.01s, ref   0.4s, 7603.52%   0.0s, -34.08%   0.0s, -80.92%  0.0s, -82.23%   0.0s, -88.21%
```

- [ ] Have you provided a meaningful PR description?
- [ ] Have you added a test, reproducer or referred to an issue with a reproducer?
- [ ] Have you tested your changes locally for CPU and GPU devices?
- [ ] Have you made sure that new changes do not introduce compiler warnings?
- [ ] If this PR is a work in progress, are you filing the PR as a draft?