@diptorupd 2023-04-08
@ZzEeKkAa I propose changing the table to something like this:

| Total data size | Platform | Benchmark | Reference | Speedup over reference | | |
| --- | --- | --- | --- | --- | --- | --- |
| | | | | numba_dpex_k| numba_dpex_p| numba_dpex_n
| 15 MB | CPU | Black-scholes | NumPy | 0.66 | 2 | 0.5 |


- Show just the relative speed up (`timing ref/timing specific impl`) in this table. We can have separate table for raw execution numbers.
- Add a separate legend table for platform
- Add a note to the report stating that any speed-up less than 1 is actually a slowdown.

I did notice that some of the timings are `0.0s`, why is that?