@adarshyoga 2023-04-10
> @ZzEeKkAa I propose changing the table to something like this:
> 
> Total data size	Platform	Benchmark	Reference	Speedup over reference		
> numba_dpex_k	numba_dpex_p	numba_dpex_n
> 15 MB	CPU	Black-scholes	NumPy	0.66	2	0.5
> * Show just the relative speed up (`timing ref/timing specific impl`) in this table. We can have separate table for raw execution numbers.

I think we should just output raw execution time numbers in seconds in a table. All second-order inferred statistics like speedup and efficiency can in a separate table or graphs. The key reason for this is that we use different baselines for different impls. For instance, for numba_dpex we use numba as baseline for speedup and dpcpp as baseline for eficiency. Similarly, for dpnp the baseline is numpy. It can get confusing if we are using different baselines for different impls in same table.

> * Add a separate legend table for platform
> * Add a note to the report stating that any speed-up less than 1 is actually a slowdown.
> 
> I did notice that some of the timings are `0.0s`, why is that? Also, can you change the legend table to left align?

It is because the execution times for 'S' are so small that when rounded to two decimal places they are show up as 0.0s.