@ZzEeKkAa 2023-04-10
I like the idea of moving speedup and efficiency completely to different report.
I've updated `s` to `ms` for better representation

```
Report for 2023-04-07 18:18:14 run
==================================
Legend
======
   impl_postfix                                    description
0  numba_dpex_k                         Numba-dpex with kernel
1  numba_dpex_p                         Numba-dpex with prange
2  numba_dpex_n         Numba-dpex with NumPy array expression
3          dpnp                   Data-parallel Numeric Python
4         numpy                                          Numpy
5        python                                         Python
6       numba_n                            Numba with nopython
7      numba_np             Numba with nopython, Parallel=True
8     numba_npr  Numba with nopython, Parallel=True and prange
9          sycl                      DPC++ with dpctl bindings

Summary of current implementation
=================================
  input_size          benchmark problem_preset numba_dpex_k numba_dpex_p numba_dpex_n       dpnp    numpy    python  numba_n numba_np numba_npr    sycl
0       15MB      black_scholes              S      25.42ms      14.64ms      43.03ms   109.02ms  22.11ms  364.79ms  15.59ms   1.22ms     1.2ms  0.95ms
1       80KB             dbscan              S      75.18ms      16.44ms          n/a        n/a      n/a    7.01ms  11.35ms      n/a    2.77ms  2.55ms
2        4KB             gpairs              S      50.38ms      33.22ms          n/a    86.59ms   0.64ms       n/a    0.3ms      n/a    0.16ms  0.28ms
3       94KB             kmeans              S     727.07ms     586.86ms          n/a        n/a      n/a  193.72ms   1.41ms      n/a    1.32ms  5.13ms
4      296KB                knn              S      30.47ms          n/a          n/a        n/a      n/a  3061.2ms      n/a      n/a    2.35ms  2.01ms
5        1MB            l2_norm              S      10.94ms       9.42ms          n/a  8663.88ms   0.31ms       n/a   0.24ms    0.2ms    0.06ms  0.23ms
6        8MB  pairwise_distance              S      18.31ms      11.68ms          n/a        n/a  11.26ms       n/a      n/a  20.54ms    0.99ms  0.89ms
7         0B                pca              S          n/a          n/a          n/a        n/a  31.44ms       n/a  66.35ms  24.45ms       n/a     n/a
8        7MB              rambo              S      28.42ms       13.6ms          n/a    38.81ms   5.23ms  403.13ms   3.45ms    1.0ms    0.93ms  0.62ms
```

Does it looks better?