@diptorupd 2023-04-11
> Report view:
> 
> ```
> Report for 2023-04-07 18:18:14 run
> ==================================
> Legend
> ======
>    impl_postfix                                    description
> 0  numba_dpex_k  Numba-dpex with kernel
> 1  numba_dpex_p  Numba-dpex with prange
> 2  numba_dpex_n  Numba-dpex with NumPy array expression
> 3  dpnp          Data-parallel Numeric Python
> 4  numpy         Numpy
> 5  python        Python
> 6  numba_n       Numba with nopython
> 7  numba_np      Numba with nopython, Parallel=True
> 8  numba_npr     Numba with nopython, Parallel=True and prange
> 9  sycl          DPC++ with dpctl bindings

Thanks for the alignment fix on the legends table. Can you also change the alignment of the column headers?