@diptorupd 2023-04-11
> Thanks for the alignment fix on the legends table. Can you also change the alignment of the column headers?
Let us merge the PR as-is and fix the formatting later.